% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunction.R
\name{simp_lin_R}
\alias{simp_lin_R}
\title{Fit a simple linear regression model}
\usage{
simp_lin_R(x, y)
}
\arguments{
\item{x}{vector, the predictor variable in simple linear regression}

\item{y}{vector, the response variable in simple linear regression}
}
\value{
a list of regression result
}
\description{
Fit a simple linear regression model
}
